Initial framework for gathering data on Items
- Moved Shared classes to separate class files: Item,
  DynamicAssetArray.Extensions, Reward
- Corrected comment hyperlinks in Component.cs
- Created initial Catalog.cs
- Corrected an exception message and added a test for a null Component
  name in Game.cs
