+++ src/classes/Game.cs
@@ -117,6 +117,7 @@ namespace Mffer {
 				Name = "";
 				Components = new Dictionary<string, Component>();
 				Data = null;
+				AddComponent( new Catalog() );
 				AddComponent( new Localization() );
 				AddComponent( new Roster() );
 			}
@@ -131,7 +132,7 @@ namespace Mffer {
 			}
 			dynamic GetData( string objectName, string dataFile ) {
 				if ( dataFile is null ) {
-					throw new ArgumentNullException( "assetFile" );
+					throw new ArgumentNullException( "dataFile" );
 				}
 				if ( !Data.DataFiles.ContainsKey( dataFile ) ) {
 					throw new KeyNotFoundException( $"Unable to find asset file named {dataFile}" );
@@ -188,6 +189,7 @@ namespace Mffer {
 			/// <param name="component"><see cref="Component"/> to associate with
 			/// this <see cref="Version"/></param>
 			void AddComponent( Component component ) {
+				if ( String.IsNullOrEmpty( component.Name ) ) throw new Exception( "The name of the component must be set" );
 				Components.Add( component.Name, component );
 			}
 			/// <summary>
