+++ src/classes/Item.cs
@@ -1,0 +1,11 @@
+namespace Mffer {
+	/// <summary>
+	/// Represents a single item
+	/// </summary>
+	public class Item : GameObject {
+		/// <summary>
+		/// Gets or sets the name of the <see cref="Item"/>
+		/// </summary>
+		public string Name { get; set; }
+	}
+}
